+++ socorro/lib/libsocorrodataschema.py
@@ -3,7 +3,7 @@
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 """
-Utilities for working with JSON schemas and JSON.
+Utilities for working with socorro-data schemas.
 """
 
 import copy
@@ -16,10 +16,6 @@ class InvalidDocumentError(Exception):
 
 class InvalidSchemaError(Exception):
     """Raised when the schema is invalid"""
-
-
-class UnknownConvertFormat(Exception):
-    """Raised when the schema specifies an unknown convert format"""
 
 
 def listify(item):
@@ -64,6 +60,15 @@ def lookup_definition(schema, ref):
 
 
 def resolve_reference(root_schema, schema, ref):
+    """Given a root schema and a schema with a ref, resolves the ref
+
+    :arg root schema: the root schema holding a "definitions" section
+    :arg schema: the schema to resolve the reference in
+    :arg ref: the ``$ref`` value
+
+    :returns: the new schema with the reference resolved
+
+    """
     # Resolve by finding the referenced schema, update it with schema data, and then
     # returning it
     #
@@ -80,10 +85,12 @@ def resolve_reference(root_schema, schema, ref):
     return new_schema
 
 
+# FIXME(willkg): rewrite as a transform
 def resolve_references(schema):
     """Returns a schema with all ``$ref`` resolved
 
-    This works for both jsonschema schemas and socorro-data schemas.
+    This works for both jsonschema schemas and socorro-data schemas because they
+    both support ``$ref`` in the same way Socorro uses it.
 
     .. Note::
 
@@ -197,7 +204,7 @@ def permissions_transform_function(permissions_have):
     return _permissions_predicate
 
 
-class FlattenSchemaKeys:
+class FlattenKeys:
     def __init__(self):
         self.keys = []
 
@@ -209,7 +216,7 @@ class FlattenSchemaKeys:
         return schema
 
 
-def transform_socorro_data_schema(schema, transform_function):
+def transform_schema(schema, transform_function):
     """Transforms a socorro-data schema using some transform_function called on nodes
 
     The ``transform_function`` gets called pre-order as the schema is traversed.
@@ -289,23 +296,6 @@ def transform_socorro_data_schema(schema, transform_function):
     return new_schema
 
 
-def convert_to(document_part, target_format):
-    """Conversion for handling historical schema goofs"""
-    # We never convert nulls
-    if document_part is None:
-        return document_part
-
-    if target_format == "string":
-        # Boolean gets convert to "0" or "1"
-        if document_part is True:
-            return "1"
-        elif document_part is False:
-            return "0"
-        return str(document_part)
-
-    raise UnknownConvertFormat(f"{target_format!r} is an unknown format")
-
-
 # Cache of pattern -> re object
 PATTERN_CACHE = {}
 
@@ -323,213 +313,6 @@ def compile_pattern_re(pattern):
     if pattern not in PATTERN_CACHE:
         PATTERN_CACHE[pattern] = re.compile(pattern)
     return PATTERN_CACHE[pattern]
-
-
-class JsonSchemaReducer:
-    """Reducer for reducing a document to the structure of the specified jsonschema
-
-    Several things to know about the reducer:
-
-    1. It doesn't support full jsonschema spec--it only supports the parts we
-       needed for our schemas.
-    2. It does some light type validation and raises exceptions for documents that
-       are invalid.
-
-    """
-
-    def __init__(self, schema):
-        """
-        :arg schema: the schema document specifying the structure to traverse
-        """
-        self.schema = schema
-
-        self._pattern_cache = {}
-
-    def get_schema_property(self, schema_part, name):
-        """Returns the matching property or patternProperties value
-
-        :arg dict schema_part: the part of the schema we're looking at now
-        :arg string name: the name we're looking for
-
-        :returns: property schema or None
-
-        """
-        properties = schema_part.get("properties", {})
-        if name in properties:
-            return schema_part["properties"][name]
-
-        pattern_properties = schema_part.get("patternProperties", {})
-        for pattern, property_schema in pattern_properties.items():
-            pattern_re = compile_pattern_re(pattern)
-            if pattern_re.match(name):
-                return property_schema
-
-        return None
-
-    def traverse(self, document):
-        """Following the schema, traverses the document
-
-        This validates types and some type-related restrictions while traversing the
-        document using the structure specified in the schema.
-
-        :arg dict document: the document to traverse using the structure specified
-            in the schema
-
-        :returns: new document
-
-        """
-        return self._traverse(schema_part=self.schema, document_part=document)
-
-    def _traverse(self, schema_part, document_part, path="", general_path=""):
-        """Following the schema, traverses the document
-
-        This validates types and some type-related restrictions while reducing the
-        document to the schema.
-
-        :arg dict schema_part: the part of the schema we're looking at now
-        :arg dict document_part: the part of the document we're looking at now
-        :arg string path: the path in the structure
-        :arg string general_path: the generalized path in the structure where array
-            indexes are replaced with `[]`
-
-        :returns: new document
-
-        """
-        # Telemetry ingestion is marred by historical fun-ness, so we first look at the
-        # schema and if it defines a converter, we run the document part through that
-        # first.
-        if "socorroConvertTo" in schema_part:
-            document_part = convert_to(document_part, schema_part["socorroConvertTo"])
-
-        schema_part_types = listify(schema_part.get("type", "string"))
-
-        # FIXME(willkg): maybe implement:
-        #
-        # * string: minLength, maxLength, pattern
-        # * integer/number: minimum, maximum, exclusiveMaximum, exclusiveMinimum,
-        #   multipleOf
-        # * array: maxItems, minItems, uniqueItems
-        # * object: additionalProperties, minProperties, maxProperties, dependencies,
-        #   regexp
-
-        # If the document_part is a basic type (string, number, etc) and it matches
-        # what's in the schema, then return it so it's included in the reduced document
-        if isinstance(document_part, BASIC_TYPES_KEYS):
-            valid_schema_part_type = BASIC_TYPES[type(document_part)]
-            if valid_schema_part_type in schema_part_types:
-                return document_part
-
-            raise InvalidDocumentError(
-                f"invalid: {path}: type {valid_schema_part_type} not in "
-                + f"{schema_part_types}"
-            )
-
-        if isinstance(document_part, list):
-            if "array" not in schema_part_types:
-                raise InvalidDocumentError(
-                    f"invalid: {path}: type array not in {schema_part_types}"
-                )
-
-            schema_items = schema_part.get("items", {"type": "string"})
-            if isinstance(schema_items, list):
-                # If schema_items is a list, then it's denoting a record like thing and the
-                # document must match the schema items in the list (or fewer)
-                new_doc = []
-                if len(document_part) > len(schema_items):
-                    raise InvalidDocumentError(f"invalid: {path}: too many items")
-
-                for i in range(0, len(document_part)):
-                    schema_item = schema_items[i]
-
-                    schema_item_path = f"{path}.[{i}]"
-                    # Since this is a record and position matters, we keep the
-                    # index number
-                    schema_item_general_path = f"{general_path}.[{i}]"
-
-                    if "$ref" in schema_item:
-                        schema_item = resolve_reference(
-                            root_schema=self.schema,
-                            schema=schema_item,
-                            ref=schema_item["$ref"],
-                        )
-
-                    new_part = self._traverse(
-                        schema_part=schema_items[i],
-                        document_part=document_part[i],
-                        path=schema_item_path,
-                        general_path=schema_item_general_path,
-                    )
-                    new_doc.append(new_part)
-                return new_doc
-
-            # If schema_items is not a list, then each document_part must match this
-            # schema
-            schema_item = schema_items
-            schema_item_path = f"{path}.[]"
-            schema_item_general_path = f"{path}.[]"
-
-            if "$ref" in schema_item:
-                schema_item = resolve_reference(
-                    root_schema=self.schema,
-                    schema=schema_item,
-                    ref=schema_item["$ref"],
-                )
-
-            new_doc = []
-            for i in range(0, len(document_part)):
-                schema_item_path = f"{path}.[{i}]"
-                schema_item_general_path = f"{path}.[]"
-                new_part = self._traverse(
-                    schema_part=schema_item,
-                    document_part=document_part[i],
-                    path=schema_item_path,
-                    general_path=schema_item_general_path,
-                )
-                new_doc.append(new_part)
-            return new_doc
-
-        if isinstance(document_part, dict):
-            if "object" not in schema_part_types:
-                raise InvalidDocumentError(
-                    f"invalid: {path}: type object not in {schema_part_types}"
-                )
-
-            new_doc = {}
-            for name, document_property in document_part.items():
-                schema_property = self.get_schema_property(schema_part, name)
-
-                # If the item is in the document, but not in the schema, we don't add
-                # this part of the document to the new document
-                if schema_property is None:
-                    continue
-
-                path_name = f"{path}.{name}"
-                general_path_name = f"{general_path}.{name}"
-
-                # Expand references in the schema property
-                if "$ref" in schema_property:
-                    schema_property = resolve_reference(
-                        root_schema=self.schema,
-                        schema=schema_property,
-                        ref=schema_property["$ref"],
-                    )
-
-                new_doc[name] = self._traverse(
-                    schema_part=schema_property,
-                    document_part=document_part[name],
-                    path=path_name,
-                    general_path=general_path_name,
-                )
-
-            # Verify all required properties exist in the document
-            for required_property in schema_part.get("required", []):
-                if required_property not in new_doc:
-                    required_property_path = ".".join([path, required_property])
-                    raise InvalidDocumentError(
-                        f"invalid: {required_property_path}: required, but missing"
-                    )
-
-            return new_doc
 
 
 class SocorroDataReducer:
